Rename file to match class and add solve_next_move assertions
